+++ CHANGELOG.md
@@ -43,6 +43,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Improvements
 
+- [#725](https://github.com/alleslabs/celatone-frontend/pull/725) Adjust unsupported asset display in account and contract detail
 - [#713](https://github.com/alleslabs/celatone-frontend/pull/713) Adjust search state consistency
 - [#712](https://github.com/alleslabs/celatone-frontend/pull/712) api v1 - proposal list
 - [#714](https://github.com/alleslabs/celatone-frontend/pull/714) api v1 - recent contracts list